Rebase structural
